requirements.txt
certifi==2019.3.9
chardet==3.0.4
coverage==4.5.3
Django==1.11.20
django-crispy-forms==1.7.2
django-faker==0.2
fake-factory==9999.9.9
Faker==1.0.7
idna==2.8
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
six==1.12.0
stripe==2.29.0
text-unidecode==1.2
urllib3==1.25.3
